// frontend/src/types/generated/FilterConstraint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EventVariable } from "./EventVariable";
import type { ObjectVariable } from "./ObjectVariable";

export type FilterConstraint =
  | {
      ObjectAssociatedWithEvent: [ObjectVariable, EventVariable, string | null];
    }
  | {
      ObjectAssociatedWithObject: [
        ObjectVariable,
        ObjectVariable,
        string | null,
      ];
    }
  | {
      TimeBetweenEvents: [
        EventVariable,
        EventVariable,
        [number | null, number | null],
      ];
    };